+++ Source/Tests/LinuxSyscalls/x64/Memory.cpp
@@ -4,6 +4,7 @@ tags: LinuxSyscalls|syscalls-x86-64
 $end_info$
 */
 
+#include "Tests/LinuxSyscalls/LinuxAllocator.h"
 #include "Tests/LinuxSyscalls/Syscalls.h"
 #include "Tests/LinuxSyscalls/x64/Syscalls.h"
 #include <FEXCore/Core/Context.h>
@@ -29,7 +30,20 @@ static std::string get_fdpath(int fd)
 namespace FEX::HLE::x64 {
   void RegisterMemory(FEX::HLE::SyscallHandler *const Handler) {
     REGISTER_SYSCALL_IMPL_X64(munmap, [](FEXCore::Core::CpuStateFrame *Frame, void *addr, size_t length) -> uint64_t {
-      uint64_t Result = FEXCore::Allocator::munmap(addr, length);
+      uint64_t Result{};
+      if (addr < (void*)0x1'0000'0000ULL) {
+        Result = (uint64_t)static_cast<FEX::HLE::SyscallHandler*>(FEX::HLE::_SyscallHandler)->Get32BitAllocator()->
+          munmap(addr, length);
+        return Result;
+
+        if (Result < -4096) {
+          errno = -Result;
+          Result = -1;
+        }
+      }
+      else {
+        Result = FEXCore::Allocator::munmap(addr, length);
+      }
 
       auto Thread = Frame->Thread;
       if (Result != -1) {
@@ -42,31 +56,24 @@ namespace FEX::HLE::x64 {
     REGISTER_SYSCALL_IMPL_X64(mmap, [](FEXCore::Core::CpuStateFrame *Frame, void *addr, size_t length, int prot, int flags, int fd, off_t offset) -> uint64_t {
       static FEX_CONFIG_OPT(AOTIRLoad, AOTIRLOAD);
 
-#ifdef _M_ARM_64
-      constexpr uint32_t X86_64_MAP_32BIT = 0x40;
+      uint64_t Result{};
 
-      // Handle MAP_32BIT without a full allocator
-      // Set address to a low address to force the kernel to allocate bottom up
-      bool Map32Bit = (flags & X86_64_MAP_32BIT) &&
-          !(flags & (MAP_FIXED | MAP_FIXED_NOREPLACE));
+      bool Map32Bit = flags & FEX::HLE::X86_64_MAP_32BIT;
       if (Map32Bit) {
-        addr = reinterpret_cast<void*>(0x1'0000);
-        // Remove the flag
-        flags &= ~X86_64_MAP_32BIT;
+        Result = (uint64_t)static_cast<FEX::HLE::SyscallHandler*>(FEX::HLE::_SyscallHandler)->Get32BitAllocator()->
+          mmap(reinterpret_cast<void*>(addr), length, prot,flags, fd, offset);
+        if (Result < -4096) {
+          errno = -Result;
+          Result = -1;
+        }
       }
-#endif
-      uint64_t Result = reinterpret_cast<uint64_t>(FEXCore::Allocator::mmap(addr, length, prot, flags, fd, offset));
+      else
+      {
+        Result = reinterpret_cast<uint64_t>(FEXCore::Allocator::mmap(addr, length, prot, flags, fd, offset));
+      }
 
       auto Thread = Frame->Thread;
       if (Result != -1) {
-#ifdef _M_ARM_64
-        if (Map32Bit) {
-          if ((Result + length) >= 0x1'0000'0000ULL) {
-            FEXCore::Allocator::munmap(reinterpret_cast<void*>(Result), length);
-            return -ENOMEM;
-          }
-        }
-#endif
         if (!(flags & MAP_ANONYMOUS)) {
           auto filename = get_fdpath(fd);
